DIT-2521 | Refactored existing reset code
