+++ app/uk/gov/hmrc/bindingtariffadminfrontend/service/DataMigrationService.scala
@@ -31,7 +31,6 @@ import play.api.libs.Files.TemporaryFile
 import uk.gov.hmrc.bindingtariffadminfrontend.connector._
 import uk.gov.hmrc.bindingtariffadminfrontend.lock.MigrationLock
 import uk.gov.hmrc.bindingtariffadminfrontend.model.MigrationStatus.MigrationStatus
-import uk.gov.hmrc.bindingtariffadminfrontend.model.Store.Store
 import uk.gov.hmrc.bindingtariffadminfrontend.model.classification.{Attachment, Case, Sample}
 import uk.gov.hmrc.bindingtariffadminfrontend.model.filestore.{FileSearch, FileUploaded}
 import uk.gov.hmrc.bindingtariffadminfrontend.model.{MigrationStatus, _}
@@ -129,28 +128,6 @@ class DataMigrationService @Inject() (
 
   def clear(status: Option[MigrationStatus] = None): Future[Boolean] =
     migrationRepository.delete(status)
-
-  def resetEnvironment(stores: Set[Store])(implicit hc: HeaderCarrier): Future[Unit] = {
-
-    def resetIfPresent(store: Store, expression: => Future[Any]): Future[Unit] =
-      if (stores.contains(store)) {
-        expression.map(_ => ()) recover loggingAWarning
-      } else Future.successful(())
-
-    def loggingAWarning: PartialFunction[Throwable, Unit] = {
-      case t: Throwable => Logger.warn("Failed to clear Service", t)
-    }
-
-    for {
-      _ <- resetIfPresent(Store.FILES, fileConnector.delete())
-      _ <- resetIfPresent(Store.FILES, uploadRepository.deleteAll())
-      _ <- resetIfPresent(Store.CASES, caseConnector.deleteCases())
-      _ <- resetIfPresent(Store.EVENTS, caseConnector.deleteEvents())
-      _ <- resetIfPresent(Store.RULINGS, rulingConnector.delete())
-      _ <- resetIfPresent(Store.HISTORIC_DATA, dataMigrationConnector.deleteHistoricData())
-      _ <- resetIfPresent(Store.MIGRATION, clear())
-    } yield ()
-  }
 
   def upload(upload: Upload, file: TemporaryFile)(implicit hc: HeaderCarrier): Future[Unit] =
     for {
