+++ app/uk/gov/hmrc/bindingtariffadminfrontend/views/data_migration_state.scala.html
@@ -43,7 +43,7 @@
     </div>
     @if(appConfig.resetPermitted) {
      <div class="inline-block">
-      <a class="button button-warn inline-block" href="@routes.DataMigrationStateController.reset()">Reset Services</a>
+      <a class="button button-warn inline-block" href="@routes.ResetController.reset()">Reset Services</a>
      </div>
     }
    </div>
