+++ app/uk/gov/hmrc/bindingtariffadminfrontend/views/index.scala.html
@@ -60,7 +60,7 @@
 
   @if(appConfig.resetPermitted){
    <li>
-    <a id="index-reset" href="@routes.DataMigrationStateController.reset()">Reset Binding Tariff services</a>
+    <a id="index-reset" href="@routes.ResetController.reset()">Reset Binding Tariff services</a>
    </li>
   }
  </ul>
